+++ util/src/sync.rs
@@ -10,11 +10,8 @@ pub mod once;
 pub mod spin;
 pub use self::once::{InitOnce, Lazy};
 
-use core::{
-    fmt,
-    ops::{Deref, DerefMut},
-};
-
+mod cache_pad;
+pub use self::cache_pad::CachePadded;
 pub mod hint {
     #[cfg(not(loom))]
     pub use core::hint::spin_loop;
@@ -29,14 +26,6 @@ pub(crate) struct Backoff {
     exp: u8,
     max: u8,
 }
-
-#[cfg_attr(any(target_arch = "x86_64", target_arch = "aarch64"), repr(align(128)))]
-#[cfg_attr(
-    not(any(target_arch = "x86_64", target_arch = "aarch64")),
-    repr(align(64))
-)]
-#[derive(Clone, Copy, Default, Hash, PartialEq, Eq)]
-pub struct CachePadded<T>(T);
 
 // === impl Backoff ===
 
@@ -74,39 +63,5 @@ impl Backoff {
 impl Default for Backoff {
     fn default() -> Self {
         Self::new()
-    }
-}
-
-// === impl CachePadded ===
-
-impl<T> CachePadded<T> {
-    pub const fn new(value: T) -> Self {
-        Self(value)
-    }
-
-    pub fn into_inner(self) -> T {
-        self.0
-    }
-}
-
-impl<T> Deref for CachePadded<T> {
-    type Target = T;
-
-    #[inline]
-    fn deref(&self) -> &T {
-        &self.0
-    }
-}
-
-impl<T> DerefMut for CachePadded<T> {
-    #[inline]
-    fn deref_mut(&mut self) -> &mut T {
-        &mut self.0
-    }
-}
-
-impl<T: fmt::Debug> fmt::Debug for CachePadded<T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
     }
 }
